+++ comb_spec_searcher/comb_spec_searcher.py
@@ -1,12 +1,22 @@
 """A class for automatically performing combinatorial exploration."""
 import gc
-import json
 import logging
 import platform
 import time
 import warnings
 from collections import defaultdict
-from typing import Any, Dict, Generic, Iterator, Optional, Sequence, Set, Tuple, cast
+from typing import (
+    Any,
+    Dict,
+    Generic,
+    Iterator,
+    Optional,
+    Sequence,
+    Set,
+    Tuple,
+    Union,
+    cast,
+)
 
 import logzero
 from logzero import logger
@@ -21,7 +31,8 @@ from .exception import (
     SpecificationNotFound,
     StrategyDoesNotApply,
 )
-from .rule_db import RuleDB, RuleDBBase, RuleDBForgetStrategy
+from .rule_db import RuleDB, RuleDBForgetStrategy
+from .rule_db.base import RuleDBBase
 from .specification import CombinatorialSpecification
 from .strategies import (
     AbstractStrategy,
@@ -58,7 +69,11 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
     """
 
     def __init__(
-        self, start_class: CombinatorialClassType, strategy_pack: StrategyPack, **kwargs
+        self,
+        start_class: CombinatorialClassType,
+        strategy_pack: StrategyPack,
+        ruledb: Optional[Union[str, RuleDB]] = None,
+        **kwargs,
     ):
         """
         Initialise CombinatorialSpecificationSearcher.
@@ -84,12 +99,16 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
         self.classdb = ClassDB[CombinatorialClassType](type(start_class))
         self.classqueue = DefaultQueue(strategy_pack)
 
-        if kwargs.get("ruledb") is None:
+        if ruledb is None:
             self.ruledb: RuleDBBase = RuleDB()
-        elif kwargs.get("ruledb") == "forget":
+        elif ruledb == "forget":
             self.ruledb = RuleDBForgetStrategy(self.classdb, self.strategy_pack)
+        elif isinstance(ruledb, RuleDBBase):
+            self.ruledb = ruledb
         else:
-            raise ValueError("ruledb argument should be None or 'forget'")
+            raise ValueError(
+                "ruledb argument should be None or 'forget' or a RuleDB object"
+            )
 
         # initialise the run with start_class
         self.start_label = self.classdb.get_label(start_class)
@@ -205,7 +224,7 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
                 children = rule.children
             except StrategyDoesNotApply:
                 continue
-            if len(children) == 1 and comb_class == children[0]:
+            if len(children) == 1 and rule.comb_class == children[0]:
                 logger.debug(
                     "The equivalence strategy %s returned the same "
                     "combinatorial class when applied to %r",
@@ -214,13 +233,18 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
                     extra=self.logger_kwargs,
                 )
                 continue
-            end_labels = [self.classdb.get_label(comb_class) for comb_class in children]
+            end_labels = [self.classdb.get_label(child) for child in children]
+            if rule.comb_class == comb_class:
+                start_label = label
+            else:
+                start_label = self.classdb.get_label(rule.comb_class)
+
             # TODO: observe that creating this constructor could be costly,
             # e.g. Cartesian
             if self.debug:
                 logger.debug(
                     "Adding combinatorial rule %s -> %s\n%s",
-                    label,
+                    start_label,
                     tuple(end_labels),
                     rule,
                     extra=self.logger_kwargs,
@@ -237,18 +261,20 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
                         "NotImplementedError: %s",
                         e,
                     )
-            if any(self.ruledb.are_equivalent(label, l) for l in end_labels):
+            if any(
+                self.ruledb.are_equivalent(start_label, elabel) for elabel in end_labels
+            ):
                 # This says comb_class = comb_class, so we skip it, but mark
                 # every other class as empty.
-                for l in end_labels:
-                    if not self.ruledb.are_equivalent(label, l):
-                        self._add_empty_rule(l)
+                for elabel in end_labels:
+                    if not self.ruledb.are_equivalent(start_label, elabel):
+                        self._add_empty_rule(elabel)
                 if self.debug:
-                    for l, c in zip(end_labels, children):
-                        if not self.ruledb.are_equivalent(label, l):
-                            assert c.is_empty()
+                    for elabel, child in zip(end_labels, children):
+                        if not self.ruledb.are_equivalent(start_label, elabel):
+                            assert child.is_empty()
 
-            yield label, tuple(end_labels), rule
+            yield start_label, tuple(end_labels), rule
 
     def _add_rule(
         self, start_label: int, end_labels: Tuple[int, ...], rule: AbstractRule
@@ -285,6 +311,9 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
             self.try_verify(comb_class, child_label)
 
             cleaned_end_labels.append(child_label)
+
+        if cleaned_end_labels == [start_label]:
+            return
 
         if rule.ignore_parent:
             self._stop_yielding(start_label)
@@ -360,7 +389,7 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
         eqs = set()
         for start, ends, strategy in self.ruledb.all_rules():
             parent = self.classdb.get_class(start)
-            children = tuple(self.classdb.get_class(l) for l in ends)
+            children = tuple(map(self.classdb.get_class, ends))
             rule = strategy(parent, children)
             try:
                 eq = rule.get_equation(get_function)
@@ -381,19 +410,15 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
             comb_class = self.classdb.get_class(label)
             self._expand(comb_class, label, strategies, inferral)
 
-    @cssiteratortimer("queue")
     def _labels_to_expand(self) -> Iterator[WorkPacket]:
         yield from self.classqueue
 
-    @cssmethodtimer("queue")
     def _add_to_queue(self, label: int):
         self.classqueue.add(label)
 
-    @cssmethodtimer("queue")
     def _not_inferrable(self, label: int):
         self.classqueue.set_not_inferrable(label)
 
-    @cssmethodtimer("queue")
     def _stop_yielding(self, label: int):
         self.classqueue.set_stop_yielding(label)
 
@@ -491,7 +516,6 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
         )
         found_string += self.status(elaborate=True)
         found_string += str(specification)
-        found_string += json.dumps(specification.to_jsonable(), separators=(",", ":"))
         logger.info(found_string, extra=self.logger_kwargs)
 
     def _log_status(self, start_time: float, status_update: int) -> None:
@@ -537,6 +561,9 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
 
         If 'smallest' is set to 'True' then the searcher will return a proof
         tree that is as small as possible.
+
+        If 'expand_verified' is set to 'False' then the searcher will not
+        expand verified classes.
         """
         auto_search_start = time.time()
 
@@ -552,7 +579,6 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
             perc = 1
         status_update = kwargs.get("status_update", None)
         max_time = kwargs.get("max_time", None)
-        smallest = kwargs.get("smallest", False)
         status_start = time.time()
         start_string = "Auto search started {}\n".format(
             time.strftime("%a, %d %b %Y %H:%M:%S", time.gmtime())
@@ -584,7 +610,11 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
                 logger.info("No more classes to expand.", extra=self.logger_kwargs)
             spec_search_start = time.time()
             logger.debug("Searching for specification.", extra=self.logger_kwargs)
-            specification = self.get_specification(smallest=smallest)
+            specification = self.get_specification(
+                smallest=kwargs.get("smallest", False),
+                expand_verified=kwargs.get("expand_verified", True),
+                minimization_time_limit=0.01 * (time.time() - auto_search_start),
+            )
             if specification is not None:
                 self._log_spec_found(specification, auto_search_start)
                 return specification
@@ -607,10 +637,15 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
 
     @cssmethodtimer("get specification")
     def get_specification(
-        self, smallest: bool = False
+        self,
+        minimization_time_limit: float = 10,
+        smallest: bool = False,
+        expand_verified: bool = True,
     ) -> Optional[CombinatorialSpecification]:
         """
         Return a CombinatorialSpecification if the universe contains one.
+
+        The minimization_time_limit only applies when smallest is false.
 
         The function will raise a SpecificationNotFound if no such
         CombinatorialSpecification exists in the universe.
@@ -624,13 +659,15 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
                 )
             else:
                 rules, eqv_paths = self.ruledb.get_specification_rules(
-                    self.start_label, iterative=self.iterative
+                    self.start_label,
+                    minimization_time_limit=minimization_time_limit,
+                    iterative=self.iterative,
                 )
         except SpecificationNotFound:
             return None
         start_class = self.classdb.get_class(self.start_label)
         comb_class_eqv_paths = tuple(
-            tuple(self.classdb.get_class(l) for l in path) for path in eqv_paths
+            tuple(map(self.classdb.get_class, path)) for path in eqv_paths
         )
         logger.info(
             "Creating a specification", extra=self.logger_kwargs,
@@ -639,4 +676,5 @@ class CombinatorialSpecificationSearcher(Generic[CombinatorialClassType]):
             start_class,
             [(self.classdb.get_class(label), rule) for label, rule in rules],
             comb_class_eqv_paths,
+            expand_verified=expand_verified,
         )