+++ comb_spec_searcher/strategies/rule.py
@@ -137,6 +137,12 @@ class AbstractRule(abc.ABC, Generic[CombinatorialClassType, CombinatorialObjectT
         """
         return self._strategy.formal_step()
 
+    @abc.abstractmethod
+    def is_equivalence(self) -> bool:
+        """
+        Returns True if the rule is an equivalence.
+        """
+
     def non_empty_children(self) -> Tuple[CombinatorialClassType, ...]:
         """
         Return the tuple of non-empty combinatorial classes that are found
@@ -180,8 +186,6 @@ class AbstractRule(abc.ABC, Generic[CombinatorialClassType, CombinatorialObjectT
 
         Raise a SanityCheckFailure error if the sanity_check fails.
         """
-        if self.comb_class.extra_parameters:
-            raise NotImplementedError("sanity check only implemented in one variable")
 
         if isinstance(self, VerificationRule):
             # TODO: test more thoroughly
@@ -190,11 +194,17 @@ class AbstractRule(abc.ABC, Generic[CombinatorialClassType, CombinatorialObjectT
         def brute_force_count(
             comb_class: CombinatorialClassType, n: int, **parameters
         ) -> int:
+            assert set(comb_class.extra_parameters) == set(
+                parameters
+            ), f"{comb_class.extra_parameters, set(parameters)}"
             return len(list(comb_class.objects_of_size(n, **parameters)))
 
         def brute_force_generation(
             comb_class: CombinatorialClassType, n: int, **parameters: int
         ) -> Iterator[CombinatorialObjectType]:
+            assert set(comb_class.extra_parameters) == set(
+                parameters
+            ), f"{comb_class.extra_parameters, set(parameters)}"
             yield from comb_class.objects_of_size(n, **parameters)
 
         actual_count = brute_force_count(self.comb_class, n, **parameters)
@@ -204,7 +214,9 @@ class AbstractRule(abc.ABC, Generic[CombinatorialClassType, CombinatorialObjectT
         )
         rule_count = self.count_objects_of_size(n, **parameters)
         self.subrecs = temprec
-        params_str = ", ".join([f"n = {n}"] + [f"{p} = {v}" for p, v in parameters])
+        params_str = ", ".join(
+            [f"n = {n}"] + [f"{p} = {v}" for p, v in parameters.items()]
+        )
         if actual_count != rule_count:
             raise SanityCheckFailure(
                 f"The following rule failed sanity check:\n"
@@ -214,7 +226,13 @@ class AbstractRule(abc.ABC, Generic[CombinatorialClassType, CombinatorialObjectT
                 f"The actual count is {actual_count}.\n"
                 f"The rule count is {rule_count}.",
             )
-
+        if any(
+            comb_class.extra_parameters
+            for comb_class in [self.comb_class, *self.children]
+        ):
+            # Can't sanity check generation of objects for classes with extra
+            # TODO test more thoroughly
+            return True
         actual_objects = set(
             list(brute_force_generation(self.comb_class, n, **parameters))
         )
@@ -318,6 +336,9 @@ class Rule(AbstractRule[CombinatorialClassType, CombinatorialObjectType]):
                 raise StrategyDoesNotApply("{} does not apply".format(self.strategy))
         return self._constructor
 
+    def is_equivalence(self):
+        return len(self.non_empty_children()) == 1 and self.constructor.is_equivalence()
+
     def backward_map(
         self, objs: Tuple[Optional[CombinatorialObjectType], ...]
     ) -> CombinatorialObjectType:
@@ -342,8 +363,8 @@ class Rule(AbstractRule[CombinatorialClassType, CombinatorialObjectType]):
         created for equivalence rules.
         """
         assert (
-            self.constructor.is_equivalence()
-        ), "EquivalenceRule can only be created for equivalence rules"
+            self.is_equivalence()
+        ), f"EquivalenceRule can only be created for equivalence rules\n{self}"
         return EquivalenceRule(self)
 
     def to_reverse_rule(self) -> "ReverseRule":
@@ -352,7 +373,7 @@ class Rule(AbstractRule[CombinatorialClassType, CombinatorialObjectType]):
         created for equivalence rules.
         """
         assert (
-            len(self.children) == 1 and self.constructor.is_equivalence()
+            self.is_equivalence()
         ), "reverse rule can only be created for equivalence rules"
         return ReverseRule(self)
 
@@ -375,7 +396,10 @@ class Rule(AbstractRule[CombinatorialClassType, CombinatorialObjectType]):
             assert (
                 self.subrecs is not None
             ), "you must call the set_subrecs function first"
-            res = self.constructor.get_recurrence(self.subrecs, n, **parameters)
+            try:
+                res = self.constructor.get_recurrence(self.subrecs, n, **parameters)
+            except AssertionError:
+                raise ValueError(f"issue with rule:\n {self}")
             self.count_cache[key] = res
         #     # THE FOLLOWING CODE SNIPPET IS FOR DEBUGGING PURPOSES
         #     if self.comb_class.extra_parameters:
@@ -471,30 +495,35 @@ class Rule(AbstractRule[CombinatorialClassType, CombinatorialObjectType]):
 class EquivalenceRule(Rule[CombinatorialClassType, CombinatorialObjectType]):
     def __init__(self, rule: Rule):
         non_empty_children = rule.non_empty_children()
-        assert (
-            len(non_empty_children) == 1 and rule.constructor.is_equivalence()
-        ), "not an equivalence rule: {}".format(str(rule))
+        assert rule.is_equivalence(), "not an equivalence rule: {}".format(str(rule))
         child = non_empty_children[0]
         super().__init__(rule.strategy, rule.comb_class, (child,))
         self.child_idx = rule.children.index(child)
         self.actual_children = rule.children
-        self._constructor: Optional[Constructor] = None
+        self._constructor: Optional[DisjointUnion] = None
 
     @property
-    def constructor(
-        self,
-    ) -> Constructor[CombinatorialClassType, CombinatorialObjectType]:
+    def constructor(self,) -> DisjointUnion:
         """
         Return the constructor, that contains all the information about how to
         count/generate objects from the rule.
         """
         if self._constructor is None:
-            self._constructor = self.strategy.constructor(
-                self.comb_class, self.actual_children
+            original_constructor = cast(
+                DisjointUnion,
+                self.strategy.constructor(self.comb_class, self.actual_children),
             )
-            if self._constructor is None:
-                raise StrategyDoesNotApply("{} does not apply".format(self.strategy))
+            assert isinstance(original_constructor, DisjointUnion)
+            self._constructor = DisjointUnion(
+                self.comb_class,
+                self.children,
+                (original_constructor.extra_parameters[self.child_idx],),
+            )
         return self._constructor
+
+    @staticmethod
+    def is_equivalence() -> bool:
+        return True
 
     @property
     def formal_step(self) -> str:
@@ -529,10 +558,7 @@ class EquivalencePathRule(Rule[CombinatorialClassType, CombinatorialObjectType])
     """
 
     def __init__(self, rules: Sequence[Rule]):
-        assert all(
-            len(rule.children) == 1 and rule.constructor.is_equivalence()
-            for rule in rules
-        )
+        assert all(rule.is_equivalence() for rule in rules)
         super().__init__(rules[0].strategy, rules[0].comb_class, rules[-1].children)
         self.rules = rules
         self._constructor: Optional[DisjointUnion] = None
@@ -540,36 +566,31 @@ class EquivalencePathRule(Rule[CombinatorialClassType, CombinatorialObjectType])
     @property
     def constructor(self) -> DisjointUnion:
         if self._constructor is None:
-            if not self.comb_class.extra_parameters:
-                return DisjointUnion(self.comb_class, self.children)
             extra_parameters: Dict[str, str] = {
                 k: k for k in self.comb_class.extra_parameters
             }
             for rule in self.rules:
-                # TODO: call extra parameters on Rule class
-                if isinstance(rule, EquivalenceRule):
-                    rules_parameters = rule.strategy.extra_parameters(
-                        rule.comb_class, rule.actual_children
-                    )[rule.child_idx]
-                elif isinstance(rule, ReverseRule):
-                    forward_rule_parameters = rule.strategy.extra_parameters(
-                        rule.children[0]
-                    )[0]
-                    rules_parameters = {
-                        b: a for a, b in forward_rule_parameters.items()
-                    }
-                else:
-                    rules_parameters = rule.strategy.extra_parameters(
-                        rule.comb_class, rule.children
-                    )[0]
+                original_constructor = rule.constructor
+                assert isinstance(original_constructor, DisjointUnion)
+                rules_parameters = original_constructor.extra_parameters[0]
                 extra_parameters = {
-                    rules_parameters[parent_var]: child_var
+                    parent_var: rules_parameters[child_var]
                     for parent_var, child_var in extra_parameters.items()
+                    if child_var in rules_parameters
                 }
+            fixed_values = {
+                k: 0
+                for k in self.children[0].extra_parameters
+                if k not in extra_parameters.values()
+            }
             self._constructor = DisjointUnion(
-                self.comb_class, self.children, (extra_parameters,)
+                self.comb_class, self.children, (extra_parameters,), (fixed_values,)
             )
         return self._constructor
+
+    @staticmethod
+    def is_equivalence() -> bool:
+        return True
 
     @property
     def formal_step(self) -> str:
@@ -648,14 +669,36 @@ class ReverseRule(Rule[CombinatorialClassType, CombinatorialObjectType]):
 
     def __init__(self, rule: Rule):
         assert (
-            len(rule.children) == 1 and rule.constructor.is_equivalence
+            rule.is_equivalence()
         ), "reversing a rule only works for equivalence rules"
         super().__init__(rule.strategy, rule.children[0], (rule.comb_class,))
         self.original_rule = rule
+        self._constructor: Optional[DisjointUnion] = None
 
     @property
-    def constructor(self) -> Constructor:
-        return self.strategy.constructor(self.comb_class, self.children)
+    def constructor(self) -> DisjointUnion:
+        if self._constructor is None:
+            original_constructor = self.original_rule.constructor
+            assert isinstance(original_constructor, DisjointUnion), (
+                "reverse rule coming from non disjoint union strategy - "
+                "you'll need to update the ReverseRule constructor!"
+            )
+            flipped_extra_params = {
+                b: a for a, b in original_constructor.extra_parameters[0].items()
+            }
+            fixed_values = {
+                k: 0
+                for k in self.children[0].extra_parameters
+                if k not in flipped_extra_params.values()
+            }
+            self._constructor = DisjointUnion(
+                self.comb_class, self.children, (flipped_extra_params,), (fixed_values,)
+            )
+        return self._constructor
+
+    @staticmethod
+    def is_equivalence() -> bool:
+        return True
 
     @property
     def formal_step(self) -> str:
@@ -695,6 +738,10 @@ class VerificationRule(AbstractRule[CombinatorialClassType, CombinatorialObjectT
             "VerificationStrategy[CombinatorialClassType, CombinatorialObjectType]",
             super().strategy,
         )
+
+    @staticmethod
+    def is_equivalence() -> bool:
+        return False
 
     def count_objects_of_size(self, n: int, **parameters: int) -> int:
         key = (n,) + tuple(sorted(parameters.items()))